Adding token utility functions to separate subfolder
