+++ client/src/Actions/WithdrawActions.js
@@ -3,8 +3,8 @@ import * as types from './actionTypes';
 import * as creators from './ActionCreators/WithdrawActionCreators';
 import { pathObj } from '../Utils/pathVariables';
 
-import { tokenIsValid } from '../Utils/tokenIsValid';
-import { tokenIsNotValid } from '../Utils/tokenIsNotValid';
+import { tokenIsValid } from '../Utils/tokenUtils/tokenIsValid';
+import { tokenIsNotValid } from '../Utils/tokenUtils/tokenIsNotValid';
 
 export const makeWithdrawal = (withdraw_details, token) => dispatch => {
   dispatch(creators.withdrawingStart())
