+++ client/src/App.js
@@ -1,6 +1,7 @@
 import React from 'react'
 import { BrowserRouter as Router, Switch, Route} from 'react-router-dom';
 
+// Styling
 import "./App.scss";
 
 // Components
